Title: 10개 입력받기, Time: 160ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240316/10개 입력받기/receive-10-inputs.java
@@ -19,7 +19,7 @@ public class Main {
             cnt++;
         }
 
-        double avg = sum / cnt;
+        double avg = (double)sum / cnt;
 
         System.out.printf("%d %.1f", sum, avg);
     }
